requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
orjson==3.10.7

